backend/requirements.txt
# Django
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5

# Authentication
djangorestframework-simplejwt>=5.3.0

# Database
psycopg2-binary>=2.9.9  # For PostgreSQL in production

# Utilities
python-dateutil>=2.8.2
Pillow>=10.1.0  # Image handling
python-dotenv>=1.0.0

# API Documentation
drf-yasg>=1.21.7

# Testing
pytest>=7.4.3
pytest-django>=4.7.0

# Development
black>=23.12.0
flake8>=6.1.0
isort>=5.13.0
